- add Avatar component to render ens or gradient avatar - add tailwind-merge to deal with classes - add AddressLink component to show addresses or ens
