+++ app/components/index.ts
@@ -12,4 +12,6 @@ export * from './Navbar';
 export * from './Footer';
 export * from './BetsListPanel';
 export * from './CardBet';
-export * from './UserAvatarAddress';
+export * from './UserAvatarWithAddress';
+export * from './Avatar';
+export * from './AddressLink';
