+++ package.json
@@ -29,6 +29,7 @@
     "react": "^18",
     "react-dom": "^18",
     "recharts": "^2.12.7",
+    "tailwind-merge": "^2.5.2",
     "viem": "2.x",
     "wagmi": "^2.5.7"
   },
